Create config file - Hotfix (#107)
* Implement Config.properties file, replace constant variables

* Added the option to choose whether to host a server or not

In the config.properties file, if HOST_SERVER is set to true, then the program will create a server and connect to it. If it is set to false, the program will NOT create a server and instead use the IP address and port number listed in the config file to try to connect to a server

Next step is to create a p2p guide

* Update ClientPortal.java

* Moving Config folder to pass maven build

* Hotfixing config issues

Config.properties file must not be succ'ed in by the .exe script. Moving it to resources should fix that

Co-authored-by: Chris <elysuim_1@hotmail.com>