Add cardBoard logger to master
CardBoardLogger.bat executes the .exe file and redirects all outputs from the console to a file name GameLogs.txt, in an effort to help us find and troubleshoot all problems players run into when servers are live. This file must be in the same folder as CardBoard.exe

CardBoardLogger.lnk is a shortcut to the .bat file and can be placed anywhere. Users will no longer have to go to Client/cardboard.exe, they can simply move this shortcut anywhere on their computer to run the game.